scripts/requirements.txt
requests>=2.31.0
PyYAML>=6.0
reportlab>=4.0.0
PyMuPDF>=1.23.0